requirements.txt
audtorch
lmdb
msgpack
msgpack-numpy
librosa
torchvision
sentencepiece
tqdm
torchsummary
python-Levenshtein
joblib
pytorch-ignite
toml
imgaug
geomloss
nemo-toolkit
nemo-asr
nemo-nlp
torch-lr-finder
torchviz
fairseq
pyrubberband
torchaudio
aubio
torchnlp